+++ source/Jitter/Collision/Shapes/CapsuleShape.cs
@@ -34,24 +34,24 @@ namespace Jitter.Collision.Shapes
     /// </summary>
     public class CapsuleShape : Shape
     {
-        private float length, radius;
+        private double length, radius;
 
         /// <summary>
         /// Gets or sets the length of the capsule (exclusive the round endcaps).
         /// </summary>
-        public float Length { get { return length; } set { length = value; UpdateShape(); } }
+        public double Length { get { return length; } set { length = value; UpdateShape(); } }
 
         /// <summary>
         /// Gets or sets the radius of the endcaps.
         /// </summary>
-        public float Radius { get { return radius; } set { radius = value; UpdateShape(); } }
+        public double Radius { get { return radius; } set { radius = value; UpdateShape(); } }
 
         /// <summary>
         /// Create a new instance of the capsule.
         /// </summary>
         /// <param name="length">The length of the capsule (exclusive the round endcaps).</param>
         /// <param name="radius">The radius of the endcaps.</param>
-        public CapsuleShape(float length,float radius)
+        public CapsuleShape(double length,double radius)
         {
             this.length = length;
             this.radius = radius;
@@ -63,8 +63,8 @@ namespace Jitter.Collision.Shapes
         /// </summary>
         public override void CalculateMassInertia()
         {
-            float massSphere = (3.0f / 4.0f) * JMath.Pi * radius * radius * radius;
-            float massCylinder = JMath.Pi * radius * radius * length;
+            double massSphere = (3.0f / 4.0f) * JMath.Pi * radius * radius * radius;
+            double massCylinder = JMath.Pi * radius * radius * length;
 
             mass = massCylinder + massSphere;
 
@@ -87,7 +87,7 @@ namespace Jitter.Collision.Shapes
         /// <param name="result">The result.</param>
         public override void SupportMapping(ref JVector direction, out JVector result)
         {
-            float r = (float)Math.Sqrt(direction.X * direction.X + direction.Z * direction.Z);
+            double r = (double)Math.Sqrt(direction.X * direction.X + direction.Z * direction.Z);
 
             if (Math.Abs(direction.Y) > 0.0f)
             {
